Fix save button
When the user submits the form for creating or editing a youth profile,
the form appears to do nothing.

This was caused by:
1) The validation function always adding the 'primaryAddress' field
   into the error object. If primary address was valid, the value of
   this field was an empty object. Otherwise it contained information
   about sub-field errors.
2) The form submit handler ensuring that the callback function would
   not be called if the the error object had any content (had any
   keys).

This meant that the form would always be in an error state, but in such
a way that the user was not able to detect it, or recover from it.

I fixed this bug by creating an extra conditional that checks if the
'primaryAddress' field has any errors, and only then adds an object for
it in the errors object.