Stop non-admin users from accessing admin UI
Previously all users were able to access the admin UI. If they did not
have correct priviledges, they weren't able to fetch data.

After this change those users are no longer able to see the admin UI.
As authentication does not provide information about user
authorization, we are using a protected query to find out if the user
is authorized to use the admin.